+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - Removed unnecessary debug messages and cleaned up unused imports.
 - Reworked entire back-end to be websocket based (big change, though little impact, for now...)
 - Refactored session logic to be more modular and easier to manage.
+- Refactor coroutines to have a general utils module.